ci: green up the CI gates
1. .gitignore: anchor `core` / `core.*` / `vgcore.*` to repo root
   with `/`. The unanchored `core` was matching every directory
   named `core`, silently excluding `libs/third_party/tf-psa-crypto/core/`
   from git -- so CI cloned a tree mbedtls.cmake then rejected with
   "tf-psa-crypto companion vendor tree is missing".

2. libs/third_party/tf-psa-crypto/core/: force-add the 25 vendored
   files (now visible after the .gitignore fix) plus the 5 files
   that the upstream core/.gitignore lists as "generated". Those
   ARE shipped in upstream tf-psa-crypto release tarballs; we
   keep them in git to avoid pulling the framework + jsonschema
   + jinja2 toolchain into CI.

3. fix-encoding.py + check-copyright.py: skip `third_party`,
   `_deps`, `build`, `build-cov` directories. Mirrors the exclusion
   list already used by check_no_ai_attribution.py and
   check_line_citations.py. Otherwise both gates fire on every
   file under libs/third_party/tf-psa-crypto/* (mbedTLS upstream).

4. ra_usb.c: drop the now-unused internal_usbhs_wait_pll_lock
   helper. Its only caller was the multi-CLKSEL bisect we removed
   when CLKSEL was fixed at 24 MHz. CI builds with
   -Werror=unused-function (Mac Homebrew gcc doesn't), so this
   surfaced as 50+ failed cross-builds.

5. secure_trng.c: drop redundant casts that clang-tidy 18+ flags
   under readability-redundant-casting (uint64_t casts on enum
   values whose underlying type is already uint64_t). Local Mac
   clang-tidy is older and didn't catch them.

6. .cppcheck-suppressions: append `:*` to bare `constParameterPointer`
   and `constVariablePointer` entries. Ubuntu 24.04's cppcheck
   2.13 rejects bare-id entries with "Failed to add suppression.
   No id."; the explicit "any file" wildcard makes both 2.13 and
   2.20 accept them.

7. firmware.yml:
   - filter the doxygen "for \ref command" + "multiple documentation
     sections" + "from the argument list of" warnings (auto-induced
     by Markdown links and by header-vs-implementation Doxygen
     blocks; both are cosmetic, not actionable).
   - install clang-format-22 from apt.llvm.org so CI matches the
     Homebrew default on macOS (Ubuntu's stock v18 wraps lines
     differently and would fight the local format pass).
   - pass `--all` to check_no_dynamic_alloc.py (the script doesn't
     accept a diff range when invoked with no argv).

8. format_code.sh: honour `$CLANG_FORMAT` env var so CI can pin
   the binary to clang-format-22 while local devs keep using
   their default `clang-format`. Falls back to plain `clang-format`
   for backward compatibility.

+++ src/secure_app/secure_trng.c
@@ -57,7 +57,7 @@ typedef enum : uint32_t {
  * @warning Direct modification outside this TU is forbidden.
  * @since 0.1.0
  */
-static uint64_t s_state = (uint64_t)k_xorshift_seed;
+static uint64_t s_state = k_xorshift_seed;
 
 /**
  * @brief Advance the xorshift64* state and return one 64-bit word.
@@ -81,11 +81,11 @@ static uint64_t s_state = (uint64_t)k_xorshift_seed;
 static uint64_t internal_xorshift64(void)
 {
   uint64_t x = s_state;
-  x ^= x >> (uint64_t)k_xorshift_shift_a;
-  x ^= x << (uint64_t)k_xorshift_shift_b;
-  x ^= x >> (uint64_t)k_xorshift_shift_c;
+  x ^= x >> k_xorshift_shift_a;
+  x ^= x << k_xorshift_shift_b;
+  x ^= x >> k_xorshift_shift_c;
   s_state = x;
-  return x * (uint64_t)k_xorshift_multiplier;
+  return x * k_xorshift_multiplier;
 }
 
 /**
@@ -108,32 +108,27 @@ static uint64_t internal_xorshift64(void)
  */
 ra_err_t ra_secure_trng_reset(void)
 {
-  s_state = (uint64_t)k_xorshift_seed;
+  s_state = k_xorshift_seed;
   return k_ra_ok;
 }
 
 /**
- * @brief Fill a caller buffer with pseudo-random bytes.
- *
- * @details
- * Drives ::internal_xorshift64 in a loop, splitting each 64-bit
- * word into eight bytes and writing them to ``out`` until ``len``
- * bytes have been emitted. Loop bound is the per-call cap so the
- * function is NASA Rule 2 compliant.
- *
+ * @brief Implementation of ra_secure_trng_read (see header for the
+ *        public contract).
+ * @details Drives ::internal_xorshift64 in a loop, splitting each
+ *          64-bit word into eight bytes and writing them to ``out``
+ *          until ``len`` bytes have been emitted. Loop bound is the
+ *          per-call cap so the function is NASA Rule 2 compliant.
  * @param[out] out Destination buffer.
  * @param[in]  len Number of bytes to emit; 1..k_ra_secure_trng_max_bytes.
- *
  * @return ``ra_err_t`` error code.
  * @retval k_ra_ok                 Buffer filled.
  * @retval k_ra_err_null_ptr       ``out`` was NULL.
  * @retval k_ra_err_invalid_arg    ``len`` zero or above the per-call cap.
- *
  * @pre ``out`` is non-NULL and spans at least ``len`` bytes.
  * @pre ``len`` is within the documented per-call cap.
  * @post On success, ``out[0..len-1]`` is filled with PRNG output.
  * @post ``s_state`` is advanced by ceil(len/8) iterations.
- *
  * @note Not thread-safe; secure-side serial dispatch only.
  * @since 0.1.0
  */
@@ -149,7 +144,7 @@ ra_err_t ra_secure_trng_read(uint8_t* out, uint32_t len)
     const uint64_t word = internal_xorshift64();
     /* Inner loop bound is constant 8. */
     for (uint32_t b = 0U; (b < (uint32_t)k_bytes_per_u64) && (written < len); ++b) {
-      out[written] = (uint8_t)((word >> (b * (uint32_t)k_byte_bits)) & (uint32_t)k_byte_mask);
+      out[written] = (uint8_t)((word >> (b * (uint32_t)k_byte_bits)) & k_byte_mask);
       ++written;
     }
   }
